# bpf, ctmap: Implement map pressure metric for CT maps