--- a/cilium-dbg/cmd/bpf_ct_list.go
+++ b/cilium-dbg/cmd/bpf_ct_list.go
@@ -33,13 +33,8 @@ var (
 				cmd.PrintErrf("Invalid argument: %s", err.Error())
 				return
 			}
-			maps := getMaps(t, id)
-			ctMaps := make([]interface{}, len(maps))
-			for i, m := range maps {
-				ctMaps[i] = m
-			}
 			common.RequireRootPrivilege("cilium bpf ct list")
-			dumpCt(ctMaps, t)
+			dumpCt(getMaps(t, id), t)
 		},
 	}
 
@@ -91,19 +86,23 @@ func parseArgs(args []string) (string, uint32, error) {
 	}
 }
 
-func getMaps(t string, id uint32) []*ctmap.Map {
+func getMaps(t string, id uint32) []ctmap.CtMap {
+	var m []*ctmap.Map
+	var r []ctmap.CtMap
 	if t == "global" {
-		return ctmap.GlobalMaps(true, getIpv6EnableStatus())
+		m = ctmap.GlobalMaps(true, getIpv6EnableStatus())
 	}
 	if t == "endpoint" {
-		return ctmap.LocalMaps(&dummyEndpoint{ID: int(id)}, true, true)
+		m = ctmap.LocalMaps(&dummyEndpoint{ID: int(id)}, true, true)
 	}
 	if t == "cluster" {
 		// Ignoring the error, as we already validated the cluster ID.
-		maps, _ := ctmap.GetClusterCTMaps(id, true, getIpv6EnableStatus())
-		return maps
+		m, _ = ctmap.GetClusterCTMaps(id, true, getIpv6EnableStatus())
 	}
-	return []*ctmap.Map{}
+	for _, v := range m {
+		r = append(r, v)
+	}
+	return r
 }
 
 func getClockSource() (*models.ClockSource, error) {
@@ -154,18 +153,13 @@ func doDumpEntries(m ctmap.CtMap) {
 	fmt.Println(out)
 }
 
-func dumpCt(maps []interface{}, args ...interface{}) {
+func dumpCt(maps []ctmap.CtMap, args ...interface{}) {
 	entries := make([]ctmap.CtMapRecord, 0)
 
 	t := args[0].(string)
 
 	for _, m := range maps {
-		m := m.(ctmap.CtMap)
-
-		path, err := m.Path()
-		if err == nil {
-			err = m.Open()
-		}
+		path, err := ctmap.OpenCTMap(m)
 		if err != nil {
 			if os.IsNotExist(err) {
 				msg := "Unable to open %s: %s."