--- a/cilium-dbg/cmd/bpf_ct_list_test.go
+++ b/cilium-dbg/cmd/bpf_ct_list_test.go
@@ -74,9 +74,7 @@ type ctRecord6 struct {
 	Value ctmap.CtEntry
 }
 
-type dumpCallback func(maps []interface{}, args ...interface{})
-
-func dumpAndRead(maps []interface{}, dump dumpCallback, c *C, args ...interface{}) string {
+func dumpAndRead[T any](maps []T, dump func([]T, ...interface{}), c *C, args ...interface{}) string {
 	// dumpCt() prints to standard output. Let's redirect it to a pipe, and
 	// read the dump from there.
 	stdout := os.Stdout
@@ -106,8 +104,7 @@ func dumpAndRead(maps []interface{}, dump dumpCallback, c *C, args ...interface{
 }
 
 func (s *BPFCtListSuite) TestDumpCt4(c *C) {
-
-	ctMaps := [2]ctmap.CtMap{
+	ctMaps := []ctmap.CtMap{
 		mockmaps.NewCtMockMap(
 			[]ctmap.CtMapRecord{
 				{
@@ -130,11 +127,7 @@ func (s *BPFCtListSuite) TestDumpCt4(c *C) {
 		),
 	}
 
-	maps := make([]interface{}, len(ctMaps))
-	for i, m := range ctMaps {
-		maps[i] = m
-	}
-	rawDump := dumpAndRead(maps, dumpCt, c, "")
+	rawDump := dumpAndRead(ctMaps, dumpCt, c, "")
 
 	var ctDump []ctRecord4
 	err := json.Unmarshal([]byte(rawDump), &ctDump)
@@ -150,8 +143,7 @@ func (s *BPFCtListSuite) TestDumpCt4(c *C) {
 }
 
 func (s *BPFCtListSuite) TestDumpCt6(c *C) {
-
-	ctMaps := [2]ctmap.CtMap{
+	ctMaps := []ctmap.CtMap{
 		mockmaps.NewCtMockMap(
 			[]ctmap.CtMapRecord{
 				{
@@ -174,11 +166,7 @@ func (s *BPFCtListSuite) TestDumpCt6(c *C) {
 		),
 	}
 
-	maps := make([]interface{}, len(ctMaps))
-	for i, m := range ctMaps {
-		maps[i] = m
-	}
-	rawDump := dumpAndRead(maps, dumpCt, c, "")
+	rawDump := dumpAndRead(ctMaps, dumpCt, c, "")
 
 	var ctDump []ctRecord6
 	err := json.Unmarshal([]byte(rawDump), &ctDump)