… go.sum
… vendor/github.com/cilium/ebpf/ARCHITECTURE.md
… vendor/github.com/cilium/ebpf/CODEOWNERS
… vendor/github.com/cilium/ebpf/CONTRIBUTING.md
… vendor/github.com/cilium/ebpf/Makefile
… vendor/github.com/cilium/ebpf/README.md
… vendor/github.com/cilium/ebpf/attachtype_string.go
… vendor/github.com/cilium/ebpf/btf/core.go
… vendor/github.com/cilium/ebpf/btf/ext_info.go
… vendor/github.com/cilium/ebpf/btf/marshal.go
… vendor/github.com/cilium/ebpf/cpu.go
… vendor/github.com/cilium/ebpf/elf_reader.go
… vendor/github.com/cilium/ebpf/elf_sections.go
… vendor/github.com/cilium/ebpf/internal/sys/mapflags_string.go
… vendor/github.com/cilium/ebpf/internal/sys/syscall.go
… vendor/github.com/cilium/ebpf/internal/sys/types.go
… vendor/github.com/cilium/ebpf/internal/sysenc/buffer.go
… vendor/github.com/cilium/ebpf/internal/unix/types_linux.go
… vendor/github.com/cilium/ebpf/internal/unix/types_other.go
… vendor/github.com/cilium/ebpf/link/anchor.go
… vendor/github.com/cilium/ebpf/link/cgroup.go
… vendor/github.com/cilium/ebpf/link/link.go
… vendor/github.com/cilium/ebpf/link/program.go
… vendor/github.com/cilium/ebpf/link/query.go
… vendor/github.com/cilium/ebpf/link/syscalls.go
… vendor/github.com/cilium/ebpf/link/tcx.go
… vendor/github.com/cilium/ebpf/linker.go
… vendor/github.com/cilium/ebpf/map.go
… vendor/github.com/cilium/ebpf/marshalers.go
… vendor/github.com/cilium/ebpf/prog.go