- go.mod: Bump github/cilium/ebpf for BatchCursor support
- maps/ctmap: Reduce code duplication to open CT maps
- bpf, ctmap: Implement map pressure metric for CT maps
- maps/ctmap: Add benchmark for batch lookups

---

Main commit for convenience:

> bpf, ctmap: Implement map pressure metric for CT maps

> This commit adds the ability to publish the CT map pressure via
> cilium_bpf_map_pressure metric. Only global CT maps are supported. Per
> endpoint CT maps (local maps) support for now is deferred due to rare
> use.
> 

> This commit adds the ability to publish the CT map pressure via
> cilium_bpf_map_pressure metric. Only global CT maps are supported. Per
> endpoint CT maps (local maps) support for now is deferred due to rare
> use.
> 
> It does this by counting the number of elements in the CT maps via batch
> map lookup, which is far more efficient than doing an element-by-element
> lookup. The counting is done via a controller that runs on a
> fixed-interval.
> 
> In order to prevent large memory allocations suddenly, the CT map dump
> is chunked into 4096 buffers. This effectively amortizes the number of
> needed syscalls to C, being the number of chunks, rather than N, being
> the number of elements in the map.
> 
> If elements are deleted / added during chunking, it does not affect hash
> maps, such as the CT map. Other map types which rely on
> `generic_map_lookup_batch()` will need to consider an implementation
> similar to `DumpReliablyWithCallback()`. See [1].
> 
> Benchmark results based on chunk size:
> 
> 1) 4096 (old) vs 2048 (new)
> 
> ```
> $ benchstat old.txt new.txt
> name                old time/op    new time/op    delta
> _MapBatchLookup-16     145ms ± 2%     148ms ± 6%     ~     (p=0.247 n=10+10)
> 
> name                old alloc/op   new alloc/op   delta
> _MapBatchLookup-16    37.0MB ± 0%    36.9MB ± 0%   -0.31%  (p=0.000 n=9+9)
> 
> name                old allocs/op  new allocs/op  delta
> _MapBatchLookup-16     3.20k ± 0%     6.40k ± 0%  +99.88%  (p=0.000 n=9+9)
> ```
> 
> 2) 4096 vs 8192
> 
> ```
> $ benchstat old.txt new.txt
> name                old time/op    new time/op    delta
> _MapBatchLookup-16     145ms ± 3%     145ms ± 3%     ~     (p=0.684 n=10+10)
> 
> name                old alloc/op   new alloc/op   delta
> _MapBatchLookup-16    37.0MB ± 0%    37.3MB ± 0%   +0.74%  (p=0.000 n=9+10)
> 
> name                old allocs/op  new allocs/op  delta
> _MapBatchLookup-16     3.20k ± 0%     1.60k ± 0%  -49.94%  (p=0.000 n=9+10)
> ```
> 
> In terms of overall performance overhead with a 30s interval for the
> controller, the difference in CPU usage for a full CT map is +8.5% in
> idle usage over a 24h period. Note, this percentage is not the same as
> the canonical CPU usage percentage that is observed in tools like `top`,
> for example. Constantly dumping the full CT map every 30s to calculate
> and update the pressure metric required an extra 0.0017 seconds per
> second according to the `container_cpu_usage_seconds_total` metric.
> 
> This was calculated by having a two node cluster where one node had a
> full CT map and the other "empty". The raw data for the above metric
> were:
>   - Full CT map node:    0.0207 seconds per second
>   - "Empty" CT map node: 0.0190 seconds per second

Fixes: https://github.com/cilium/cilium/issues/20069