@christarazi 2023-11-30
@lmb It's up to date with the current version of cilium/ebpf because I thought that it wasn't specifically necessary to wait for the new changes in https://github.com/cilium/ebpf/pull/1223. Ultimately, it's just an API improvement, but nothing fundamental to this PR, IMO.